ContactList/requirements.txt
agent==0.1.2
aiohttp==3.6.2
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5
beautifulsoup4==4.9.1
boto==2.49.0
boto3==1.13.25
botocore==1.16.25
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==2.2.10
django-cors-headers==3.3.0
django-crispy-forms==1.9.1
django-model-utils==4.0.0
django-redis==4.12.1
django-redis-cache==2.1.1
django-ses==0.8.14
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-gis==0.15
docutils==0.15.2
elasticsearch==7.7.1
filelock==3.0.12
future==0.18.2
GDAL==2.1.2
get==2019.4.13
googlemaps==4.4.1
gunicorn==20.0.4
idna==2.9
isort==5.6.4
jellyfish==0.6.1
jmespath==0.10.0
lazy-object-proxy==1.4.3
lxml==4.5.1
lz4==3.0.2
mccabe==0.6.1
more-itertools==8.3.0
multidict==4.7.6
newrelic==5.14.0.142
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pathlib-mate==1.0.0
pep8==1.7.1
phonenumbers==8.12.5
Pillow==7.1.2
pluggy==0.13.1
post==2019.4.13
probableparsing==0.0.1
psycopg2-binary==2.8.5
public==2019.4.13
py==1.8.1
pycodestyle==2.5.0
pyfiglet==0.8.post1
pygame==1.9.6
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
pytest==5.4.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
query-string==2019.4.13
redis==3.5.3
request==2019.4.13
requests==2.23.0
rethinkdb==2.3.0.post6
retrying==1.3.3
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.1
twilio==6.41.0
urllib3==1.25.8
us==2.0.2
usaddress==0.5.10
uszipcode==0.2.4
virtualenv==20.0.21
virtualenvwrapper-win==1.2.6
wcwidth==0.2.4
wrapt==1.12.1
yarl==1.4.2
